Config: Move the editor setting defaults out of the language server
