+++ javascript/packages/config/src/index.ts
@@ -1,4 +1,4 @@
-export { Config, resolveSeverity, ALL_RULES_KEY } from "./config.js"
+export { Config, resolveSeverity, ALL_RULES_KEY, defaultPersonalSettings } from "./config.js"
 export { HerbConfigSchema, FRAMEWORKS, FRAMEWORK_NAMES } from "./config-schema.js"
 export { addHerbExtensionRecommendation, getExtensionsJsonRelativePath } from "./vscode.js"
 
@@ -16,7 +16,8 @@ export type {
   FromObjectOptions,
   ConfigValidationError,
   SeverityConfig,
-  LinterMode
+  LinterMode,
+  PersonalHerbSettings
 } from "./config.js"
 
 export type { VSCodeExtensionsJson } from "./vscode.js"
